Back to working state
- After moving pooltool/events/resolve.py -> pooltool/physics/engine.py,
  everything is back to a working state
