+++ pooltool/__init__.py
@@ -23,7 +23,6 @@ from pooltool.events import (
     filter_time,
     filter_type,
     null_event,
-    resolve_event,
     rolling_spinning_transition,
     rolling_stationary_transition,
     sliding_rolling_transition,
@@ -55,6 +54,7 @@ from pooltool.objects import (
     TableModelDescr,
     TableType,
 )
+from pooltool.physics.engine import resolve_event
 from pooltool.system import MultiSystem, System, SystemController, multisystem, visual
 
 run = terminal.Run()
@@ -74,7 +74,6 @@ __all__ = [
     "filter_time",
     "filter_type",
     "FrameStepper",
-    "resolve_event",
     "null_event",
     "ball_ball_collision",
     "ball_linear_cushion_collision",
